requirements.txt
google-cloud-aiplatform==1.121.0
google-cloud-bigquery==3.38.0
google-cloud-pipeline-components==2.21.0
google-cloud-storage==2.14.0
joblib==1.3.2
kfp==2.14.6
kfp-kubernetes==2.14.6
numpy==2.3.4
pandas==2.3.3
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.7.2
tqdm==4.67.1

